+++ R/process_bundles.R
@@ -69,33 +69,33 @@ dt_process_bundles <- function(bundles = character(), ...) {
             }
         })
 
-    merge <- function(x) {
-        if (length(x) > 1) {
-            x_ <- confx::conf_merge(x[[1]], x[[2]])
-            x[[2]] <- x_
-            Recall(x[-1])
-        } else {
-            x
-        }
-    }
+    # merge <- function(x) {
+    #     if (length(x) > 1) {
+    #         x_ <- confx::conf_merge(x[[1]], x[[2]])
+    #         x[[2]] <- x_
+    #         Recall(x[-1])
+    #     } else {
+    #         x
+    #     }
+    # }
 
-    postprocess <- function(.x) {
-        .x %>% purrr::imap(function(.opt, .name) {
-            if (.name == "dom") {
-                .opt[which.max(nchar(.opt))]
-            } else {
-                if (!is.list(.opt)) {
-                    unique(.opt)
-                } else {
-                    postprocess(.opt)
-                }
-            }
-        })
-    }
+    # postprocess <- function(.x) {
+    #     .x %>% purrr::imap(function(.opt, .name) {
+    #         if (.name == "dom") {
+    #             .opt[which.max(nchar(.opt))]
+    #         } else {
+    #             if (!is.list(.opt)) {
+    #                 unique(.opt)
+    #             } else {
+    #                 postprocess(.opt)
+    #             }
+    #         }
+    #     })
+    # }
 
     bundle_values %>%
         # Merge
-        merge() %>%
+        dt_process_bundles_merge() %>%
         # Flatten
         purrr::flatten() %>%
         # Postprocess
@@ -104,7 +104,7 @@ dt_process_bundles <- function(bundles = character(), ...) {
                 return(.x)
             }
 
-            .x %>% postprocess()
+            .x %>% dt_process_bundles_postprocess()
         })
 }
 
@@ -115,30 +115,6 @@ dt_process_bundles_list <- function(
     extensions <- bundles %>% purrr::map("extensions")
     options <- bundles %>% purrr::map("options")
 
-    merge <- function(x) {
-        if (length(x) > 1) {
-            x_ <- confx::conf_merge(x[[1]], x[[2]])
-            x[[2]] <- x_
-            Recall(x[-1])
-        } else {
-            x
-        }
-    }
-
-    postprocess <- function(.x) {
-        .x %>% purrr::imap(function(.opt, .name) {
-            if (.name == "dom") {
-                .opt[which.max(nchar(.opt))]
-            } else {
-                if (!is.list(.opt)) {
-                    unique(.opt)
-                } else {
-                    postprocess(.opt)
-                }
-            }
-        })
-    }
-
     # Merge extensions
     extensions <- extensions %>% unlist() %>% list(.) %>%
         purrr::set_names("extensions")
@@ -146,21 +122,85 @@ dt_process_bundles_list <- function(
     # Merge options
     options <- options %>%
         # Merge
-        merge() %>%
+        dt_process_bundles_list_merge() %>%
         # Flatten
         purrr::flatten() %>%
         # Postprocess
-        # purrr::imap(function(.x, .y) {
-        #     if (!is.list(.x)) {
-        #         return(.x)
-        #     }
-        #
-        #     .x %>% postprocess()
-        # }) %>%
-        postprocess() %>%
+        dt_process_bundles_list_postprocess() %>%
         list(.) %>%
         purrr::set_names("options")
 
     # Combine
     c(extensions, options)
+}
+
+# Utils -------------------------------------------------------------------
+
+dt_process_bundles_merge <- function(x) {
+    if (length(x) > 1) {
+        x_ <- confx::conf_merge(x[[1]], x[[2]])
+        x[[2]] <- x_
+        Recall(x[-1])
+    } else {
+        x
+    }
+}
+
+dt_process_bundles_postprocess <- function(.x) {
+    .x %>% purrr::imap(function(.opt, .name) {
+        if (.name == "dom") {
+            .opt[which.max(nchar(.opt))]
+        } else {
+            if (!is.list(.opt)) {
+                unique(.opt)
+            } else {
+                postprocess(.opt)
+            }
+        }
+    })
+}
+
+#' Align multiple `dom` inputs
+#'
+#' @param x
+#'
+#' @return
+#'
+#' @examples
+dt_process_bundles_list_postprocess_align_dom <- function(x) {
+    tmp <- x %>%
+        stringr::str_split("") %>%
+        unlist() %>%
+        unique()
+
+    c(
+        tmp %>% stringr::str_subset("[A-Z]"), #%>% sort(),
+        tmp %>% stringr::str_subset("[a-z]") #%>% sort()
+    ) %>%
+        stringr::str_c(collapse = "")
+}
+
+dt_process_bundles_list_postprocess <- function(.x) {
+    .x %>% purrr::imap(function(.opt, .name) {
+        if (.name == "dom") {
+            # .opt[which.max(nchar(.opt))]
+            .opt %>% dt_process_bundles_list_postprocess_align_dom()
+        } else {
+            if (!is.list(.opt)) {
+                unique(.opt)
+            } else {
+                dt_process_bundles_list_postprocess(.opt)
+            }
+        }
+    })
+}
+
+dt_process_bundles_list_merge <- function(x) {
+    if (length(x) > 1) {
+        x_ <- confx::conf_merge(x[[1]], x[[2]])
+        x[[2]] <- x_
+        Recall(x[-1])
+    } else {
+        x
+    }
 }